+++ src/Screens/HashVerificationPage.css
