+++ src/Screens/VotingPage.css
